Container restart metric (#9)
* Rework container restart metric

This commit reworks the container restart metric in the same manner as
deployment metrics. The format remains unchanged to the previous format
as in one metric per container via labels.

* Normalize expected string representation

To be able to compare strings easier when a failure occurs, remove empty
lines and trim whitespace for each line of the expected metrics string.