Added metrics describing the cpu and mem requests per container . These metrics are retrieved using the v1 api directly. No additional acucmulation or computation is done.
The state is maintained per container in a pod in kubernetes.
This change exposes the state in kubernetes via  kube-state-metrics

The metrics are also tagged by the node,where  the containers are residing.
Our use case for these metrics are as follows:

We would like to accumulate these per node and understand how much
remaining un-requested resourced are available in each node.

This would give an indication of the "utilization" of the cluster to us.