collectors/pod: flatten created by label and extract owners
The created by label previously was in the format "<kind>/<name>", this
is now normalized in to a respective label for each.

Additionally it is possible that a Pod has multiple OwnerReferences.
Previously multiple `kube_pod_info` metrics were generated for each
OwnerReference, unnecessarily duplicating all other labels of the
`kube_pod_info` metric.