Add kube_namespace_created and kube_namespace_labels
This allows prometheus reporting based on labels per namespace.

For example, if each namespace has a team label, you can report on
the memory use per team like this:

    sum(sum(container_memory_usage_bytes) by(namespace) * on(namespace)
    group_right() kube_namespace_labels) by (label_team)

We use the label "namespace" for the name of the namespace, to make it
easy to join with other metrics which is the primary use case.