Match deployment resources to recommendation
Adjust the deployment to match the recommended resources (#196). The baseline resources are set to the basic recommendation for a 100 node cluster.

Pod nanny does not support "the baseline includes the first 100 nodes", so instead set the threshold until it needs to adjust wide. This over-estimates resource needs for intermediate cluster sizes, but I'd rather be on the safe side.