kubernetes/deployment.yaml: Remove addon resizer
The addon resizer sidecar of the kube-state-metrics deployment caused
problems recently e.g. starting multiple pods [1]. In addition it seems
to be best practice to solve auto-scaling not on a per pod sidecar
basis, but instead via the autoscaling project.

This commit removes the sidecar, leaving autoscaling up to the user.

[1] https://github.com/kubernetes/kube-state-metrics/issues/672