tests/e2e.sh: Check for errors in the logs
kube-state-metrics does not terminate the pod on errors, so to actually
detect the errors in CI, we are adding this check by grepping for errors
in the kube-state-metrics logs.

This makes the assumption that any errors that are present in logs start
wtih the string 'E<MONTHDAY>' (e.g.: 'E0711') as this is how klog starts
all error log lines.