correct mapping of hpa condition status labels
The "condition" and "status" labels for the hpa status conditions were
mapped to the incorrect values. This resulted in the status being in the
condition label, and the condition in the status label.

This changelist corrects the mapping, so that condition and status map
to their respective values.

    kube_hpa_status_condition{condition="AbleToScale",hpa="hpa1",namespace="ns1",status="false"} 0
    kube_hpa_status_condition{condition="AbleToScale",hpa="hpa1",namespace="ns1",status="true"} 1
    kube_hpa_status_condition{condition="AbleToScale",hpa="hpa1",namespace="ns1",status="unknown"} 0

Fixes: f9658ca ("Add hpa conditions")
Signed-off-by: Terin Stock <terin@cloudflare.com>
(cherry picked from commit 6b09c5eb7d671c5faa7ef91a07ebc88bfb5ad9d1)