[Pod] remove kubctl logic in kube_pod_status_phase
As explained in the doc: "kube-state-metrics exposes raw data
unmodified from the Kubernetes API".

The goal of this change was to remove the kubctl logic introduced to
get the `Unknown` phase in `kube_pod_status_phase` metric.

With the introduction of `kube_pod_deleted` metric, it is now possible
to get the same result with a promQL query. (more info in this #1013).

Signed-off-by: cedric lamoriniere <cedric.lamoriniere@datadoghq.com>