Remove Lease from test manifests
The coordination.k8s.io/Lease used in e2e tests had an ownerReference,
referencing non-existing node "kube-master" (non-existing in default
Minikube deployments), so the lease was never created, and there was
an (ignored) error during deletion.

But, as a lease is automatically created for each node, there will always
be leases in e2e, so there's no need to add one.