Avoid conflicts when mapping Kubernetes labels to Prometheus labels
Prometheus 2.16.0 introduced a breaking change that prevents
_any_ metrics from being ingested from a target in case
_even a single_ metric has duplicate labels.

See https://github.com/prometheus/prometheus/pull/6664

This becomes a problem for ksm as Kubernetes label naming is less restrictive
than Prometheus label naming, so we have a "sanitization" process that replaces
unsupported characters with an underscore.

A side effect of this process is that it can produce duplicate labels, e.g.
Kubernetes labels "foo.bar", "foo-bar" and "foo_bar" are mapped to the same
Prometheus label "foo_bar".

To work around this, we're only adding the first label (sorted by `sort.Strings()`)
and discarding the rest of duplicates. In the example above, the value of
"foo-bar" will be used because hyphen (45) comes before dot (46) and underscore (95)
in ASCII.

If you'd like to have more control over which value is chosen,
you should consider addressing this on a different level of the stack,
e.g. by standardizing Kubernetes labels using Admission Webhook.