feat(persistentvolume): introduce kube_persistentvolume_claim_ref
Adds a new metrics to provide claimRef information to Prometheus. In
tracking down lingering PVs, it would be of interest to know where the
PV came from. With the claimRef information, it is easier to track down
the resource owners and inform that they have unclaimed PVs lingering
around.

Main motivation is that PVs cost money at cloud providers, if PVs stick
around for long, the costs might increase significantly.