[CARRY] internal/store: allow all node labels by default
Some of the Telemetry rules need specific node labels to be present to
provide the expected results. The set of allowed labels will be
configured in CMO but it can't happen before kube-state-metrics is
bumped to v2.0 (chicken and egg problem). As a temporary fix, the
downstream kube-state-metrics will expose all of the node labels if the
list of allowed labels is empty.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>