Properly initialize KSM server test
The pkg/app/server test does not properly initialize KSM with an empty
metrics opt-in list. This results in opt-in metrics showing up in the
expected test results.

This commit changes the test to be in line with how KSM is initialized
in production code by passing in an empty opt-in list.