Update VERSION to reflect the latest patch release
Reflect patch release changes on main, which was not the case with
2.10.1 since that diverged and was never merged back into main (around
backporting concerns).

I believe in the future we can merge everything
back into main to reflect the latest version changes in main, but
diverge the release branches starting from the pre-merge commit to make
sure backports are clean and maintainable (these will not be merged back
into the default branch).

Signed-off-by: Pranshu Srivastava <rexagod@gmail.com>