Add allowedLabels to all metrics instead of just `_labels` metric
KSM allows providing a list of labels that should be added to the prometheus metric for the resource.
These labels currently are only added to the `kube_<resource>_labels` metric. However, there are several
use cases where having these labels on all of the metrics would be helpful. This change adds the allowed
labels to all metrics of the resource instead of just the `_labels` metric.