Add info metric for client-go version
KSM supports a range of Kubernetes versions which makes it important to
keep up with new versions as clusters upgrade.

In order to allow for alerting on unsupported cluster versions, this commit
adds an info metric for the client-go version used by KSM. This way users can
set up alerts if the cluster in which KSM is running is outside of the supported range.

Signed-off-by: fpetkovski <filip.petkovsky@gmail.com>