feat: warn on probe failing
We have a fairly common occurence of KSM getting restarted due to its
liveness probe failing. This is seemingly caused by rolling deploys of
the backing API server, but we don't know exactly why. This should make
it slightly more visible what kind of response the API server when a
probe fails, which in turn could help either tweak probes to weather
such situations where the API server may be unavailable or overwhelmed
(it's entirely possible it's returning KSM a bunch of 429s) or if we
have a larger problem at hand.